Add ability to show recent tasks
Added a recents activity with a RecyclerView to show recent tasks and a
service to incept events from SystemUI.
Each recent task is dismissible by swiping up on the task and can be
opened by clicking on the task. The clear all button clears all tasks
from recents. Dismissing recents would open the last running task before
opening recents.

RecentTasksProvider holds logic to fetch and refresh list of recent
tasks from the system and updates them by listening to the system. It
communicates changes using RecentsDataChangeListener.

RecentTasksViewModel is the bridge between the provider and views.
RecentTasksViewModel holds logic to hide apps that are automotive
specific. It uses HiddenPackageListProvider that can be used to provide
a dynamic list of packages to be hidden from recents. It communicates
changes using RecentTasksChangeListener.

Bug: 271166940
Test: manual, atest-dev com.android.car.carlauncher

Change-Id: I58f62841f254ee1509a9998a6336af5fca6a290c