Update drag shadow animation after drop action
All drop animations are currently animated by
the current Window Session. This does not account for
cases when we do not want to use the default
return animation or drop animation, such as when we
need to animate an icon into a specific location or
disappear after a delay.

We add the permission required to access dragSurface
and request Window Manager to relinquish control over
the SurfaceControl, and perform the drag animation
an app side.

This change also adds driver safety to cancel drag
when user changes to drive mode while dragging.

Bug: 269608657
Test: manual - drag app icons around the launcher
Test: manual - change to drive mode and see dragged
icon disappear

Change-Id: I0e893d97b698fb3dad5f12f5e565b57c56e80630