Fix snapping views in Recents
Added a mechanism to correctly snap views within recents and add
appropriate padding to the RecyclerView. A page is defined by setting
the number of columns consituting a page. The snapper brings the
nearest center of view(for first view) or page of views to the center
of the screen. Padding is added in the beginning and the end of the
recyclerview to make the first view and fit the last page of views
centerd. Hidden ViewHolders are added to always complete the last
page.

Fixes: 276762377
Fixes: 273819043
Bug: 271166940
Test: Manual, added tests
Change-Id: I53fa69a11b6274658c0b5a5561de891807d477e9
Merged-In: I53fa69a11b6274658c0b5a5561de891807d477e9