Fix snapping for 3+ columns per page
While developing the feature it was decided to have fixed width of items
in the RecyclerView. The previous algorithm calculated the center of a
page using the average center of all the containing views. This failed
in situations where LayoutManager did not prepare the childern since
they were not close to screen. Hence the calculcated center was
incorrect. The new way to determine center depends on the start of the
page and the calculated page width(from the fixed width items).

Fixes: 292103487
Bug: 271166940
Test: atest CarLauncherTests
Change-Id: I9e7d46bbbe6d607dc65eb1453ba84d7c61f4f229