Add mechanism to notify Dock
Added DockEvents which consists of events that can be sent to the dock
to trigger a change.

Added a broadcast listener which would be used to listen to incoming
events. The receiver is added to the lib's Manifest so only explicit
intents can trigger the broadcast.

Added a helper for senders to be able to send correct DockEvents to the
receiver.

Bug: 301483111
Bug: 301482374
Test: atest CarDockLibTests
Change-Id: I4a25e9c7301d81fc772f824ef92ea9613017c574