Add drop animations to dock items
When dock item is dropped, a surface(drag surface with new icon) is
animated into the final place using two animations:
1. scale down animation: the surface is scaled down smaller than final
   size.
2. scale up animation: the surface is scaled up from the previous
   animation to the final size.

After this animation the surface is cleaned up(hidden and removed)
after the app icon is updated to the new item that was dropped into the
spot. This is achieved using View.post().

Bug: 312737692
Bug: 301482374
Test: atest CarDockLibTests, added video to the bug
Change-Id: Ic1dd492d9ee2e05634b78d07d9090b0b9146d32e